Fix: the stored policy was a byte array, wrapping moves above 255
With tiers and five spending levels the menu has 360 moves (432 with six), but
pol was a Uint8Array, so any stored index above 255 read back as a different
move. chooseAction reads the stored move in the final year of every simulated
path, so every full-menu simulation played its last year with the wrong move
wherever the best one was numbered above 255, and the E1 probe's persistence
figures were read from the same corrupted table. Now Uint16Array, with a test
that fails on the old width. audit-pol-overflow.mjs measures what it cost.

+++ src/solver/solve.js
@@ -397,7 +397,9 @@ export function solve(E, M, plan, opts = {}) {
   const survW = [], lsurvW = [], resilW = [], lresilW = [], beqW = [], polW = [], shortW = [];
   for (let k = 0; k < K; k++) {
     survW[k] = mk(); lsurvW[k] = mk(); resilW[k] = mk(); lresilW[k] = mk(); beqW[k] = mk(); shortW[k] = mk();
-    polW[k] = []; for (let t = 0; t <= T; t++) polW[k][t] = new Uint8Array(g.size);
+    // Uint16, not Uint8: with tiers and five spending levels the menu has 360 moves (432 with six), and a
+    // byte silently wrapped every index above 255 to a different move. Found 23 Sep; see results-pol-overflow.txt
+    polW[k] = []; for (let t = 0; t <= T; t++) polW[k][t] = new Uint16Array(g.size);
   }
   const surv = survW[centre], lsurv = lsurvW[centre], resil = resilW[centre];
   const lresil = lresilW[centre], beq = beqW[centre], pol = polW[centre], short = shortW[centre];
